Modificaciones Varias y desarrollo de Seeder

+++ app/Http/Controllers/UsuariosController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Usuarios;
 use App\Models\User;
+use App\Models\Auditoria;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Auth;
@@ -14,7 +15,7 @@ class UsuariosController extends Controller
     public function index()
     {
       
-        $datos =  Usuarios::orderBy('id', 'desc')->paginate(3);
+        $datos =  Usuarios::orderBy('id', 'desc')->paginate(10);
 
         return view('inicio', compact('datos')); //ENTRE COMILLAS SIMPLES
     }
@@ -63,6 +64,18 @@ class UsuariosController extends Controller
     $user->save();
 
     //AMBAS TABLAS TIENEN UNA RELACIÓN 1 A 1
+
+
+    /****LOGICA PARA AUDITORIAS****/
+    $auditoria = new Auditoria();
+    $auditoria->fecha_hora = now();
+    $auditoria->usuario_id = Auth::id();
+    $auditoria->accion = "Agregó a un usuario";
+    $usuarioResponsable = Usuarios::find(Auth::id());
+    $auditoria->nombre_usuario = $usuarioResponsable->apellido . ', ' . $usuarioResponsable->nombre;   
+    $auditoria->save();
+
+
 
     
 
@@ -122,6 +135,18 @@ class UsuariosController extends Controller
          if ($verificacion == $verificacionUsuario) {
             $usuarios->update();
             $user->update();
+
+             /****LOGICA PARA AUDITORIAS****/
+             $auditoria = new Auditoria();
+             $auditoria->fecha_hora = now();
+             $auditoria->usuario_id = Auth::id();
+             $auditoria->accion = "Modificó un usuario";
+             $usuarioResponsable = Usuarios::find(Auth::id());
+             $auditoria->nombre_usuario = $usuarioResponsable->apellido . ', ' . $usuarioResponsable->nombre;   
+             $auditoria->save();
+
+
+
             return redirect()->route("usuarios.index")->with("success", "Usuario modificado con éxito, no se modificó el email.");
         }
 
@@ -132,6 +157,16 @@ class UsuariosController extends Controller
 
          $usuarios->update();
          $user->update();
+
+           /****LOGICA PARA AUDITORIAS****/
+           $auditoria = new Auditoria();
+           $auditoria->fecha_hora = now();
+           $auditoria->usuario_id = Auth::id();
+           $auditoria->accion = "Modificó un usuario";
+           $usuarioResponsable = Usuarios::find(Auth::id());
+           $auditoria->nombre_usuario = $usuarioResponsable->apellido . ', ' . $usuarioResponsable->nombre;   
+           $auditoria->save();
+
          return redirect()->route("usuarios.index")->with("success", "Usuario modificado con éxito.");
 
          
@@ -140,12 +175,27 @@ class UsuariosController extends Controller
     
     public function destroy(Request $request, $id)
     {
+        $verificacionID_SesionActual = Auth::id();
+
+        if($verificacionID_SesionActual == $id){
+            return redirect()->route("usuarios.index")->with("success", "No puede eliminar su propio usuario"); 
+        }
+
         $usuarios = Usuarios::find($id);
         $user = User::find($id);
 
         //METODO PARA ELIMINAR
         $usuarios->delete();
         $user->delete();
+
+          /****LOGICA PARA AUDITORIAS****/
+          $auditoria = new Auditoria();
+          $auditoria->fecha_hora = now();
+          $auditoria->usuario_id = Auth::id();
+          $auditoria->accion = "Eliminó un usuario";
+          $usuarioResponsable = Usuarios::find(Auth::id());
+          $auditoria->nombre_usuario = $usuarioResponsable->apellido . ', ' . $usuarioResponsable->nombre;   
+          $auditoria->save();
 
         //PARA RETORNAR
         return redirect()->route("usuarios.index")->with("success", "Eliminado con éxito!");
